# fix(docs-infra): speed up ng test