@cexbrayat 2020-11-30
Strange, I'm also on MacOS 84 with headless Chome 🤔 The "real life" projects we work on saw a 2x-3x improvement as there are thousands of tests. It's also more visible on CI as I guess the memory pressure is more important on smaller machines (and maybe that's the difference between us: I have a fairly old laptop).
No need to waste too much on this, I'll close the PR ;)